(build): the editor compiles with the build's flags, and the linter means one thing everywhere (#101)
* (build): the editor compiles with the build's flags, and the linter means one thing everywhere

Configuring writes the compile database and cpp/.clangd points clangd at
it, so an editor stops guessing include roots, generated headers and the
language standard. clang-tidy arrives at the pinned LLVM version with a
curated check set in .clang-tidy, each exclusion carrying its reason,
enforced by the workflow and shown live by clangd. The findings the
first sweep produced are fixed where the advice was right and refused
inline where the copy or the seed was the point.

* (build): the comparison's header, spelled out where gcc reads it

* (build): the walk's end is borrowed, which libstdc++ prices and libc++ hid

+++ cpp/benchmarks/timings.hpp
@@ -80,8 +80,9 @@ class Timings {
       return out;
     }
     const auto at = [&values](const double percentile) {
-      const std::size_t index =
-          std::min(values.size() - 1, static_cast<std::size_t>(percentile / 100.0 * values.size()));
+      const std::size_t index = std::min(
+          values.size() - 1,
+          static_cast<std::size_t>(percentile / 100.0 * static_cast<double>(values.size())));
       return values[index];
     };
     out.p50 = at(50.0);
